The company_truck property of job_offer_data can't be empty, so get it from the server
ETS2 updated with the release of the Heavy Cargo DLC, and in some cases
those changes meant that upon syncing the property company_truck in some
job_offer_data units would be empty. If that happened and that unit had
a job set by the app, the game would reset the economy.